use hilt navigation library for creating vm
